refactor(telegram): extract body and preview helpers from streamResponse (#1073)
Extract two more concern siblings from the streaming.ts god-function, reducing
it from 1143 lines (before this PR series) to 948 lines:

- streaming.body.ts (47 LOC) — computeFinalBody(), the module-level replacement
  for the inner finalBody() closure. Takes explicit FinalBodyState instead of
  closing over mutable outer scope, making it testable in isolation.

- streaming.preview.ts (195 LOC) — computeLivePreview(), renderProgressRegion(),
  renderInterleavedPreview(), safeSplicePoint(), MAX_PROGRESS_LINES,
  MAX_PROGRESS_ENTRIES, and the ProgressEntry type. Extracted with full re-exports
  from streaming.ts so all existing callers (tests, other modules) are unaffected.

Together with the previously committed streaming.retry.ts (65 LOC) and
streaming.activity.ts (103 LOC), the streamResponse god-function's concerns are
now distributed across five focused siblings. The public surface of streaming.ts
is unchanged — all re-exports remain.

Baselines updated: streaming.ts 1143→948 LOC, streamResponse 819→811 function LOC.
