+++ .filesize-baseline.json
@@ -130,13 +130,13 @@
     "src/skills/audit-fit/index.ts": { "loc": 592, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/skills/mint/index.ts": { "loc": 443, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/skills/score/index.ts": { "loc": 390, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
-    "src/telegram/bot.ts": { "loc": 637, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
+    "src/telegram/bot.ts": { "loc": 644, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/telegram/elicitation-handler.ts": { "loc": 459, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/telegram/formatter.ts": { "loc": 587, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/telegram/handlers/farm-callbacks.ts": { "loc": 623, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/telegram/handlers/message.ts": { "loc": 955, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/telegram/session-manager.ts": { "loc": 896, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
-    "src/telegram/streaming.ts": { "loc": 1143, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
+    "src/telegram/streaming.ts": { "loc": 948, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/telegram/watch.ts": { "loc": 382, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/web-server/routes.ts": { "loc": 352, "reason": "legacy: predates the ceiling gate; pending concern extraction" }
   }
