+++ .funcsize-baseline.json
@@ -53,7 +53,7 @@
     "src/telegram/bot.ts::TelegramBot.setupHandlers": { "loc": 247, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/telegram/elicitation-handler.ts::makeTelegramElicitationHandler": { "loc": 381, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/telegram/handlers/message.ts::MessageHandler.handlePhoto": { "loc": 234, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/telegram/streaming.ts::streamResponse": { "loc": 819, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/telegram/streaming.ts::streamResponse": { "loc": 811, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/web-server/frontend/view-model.ts::createTranscriptModel": { "loc": 208, "reason": "legacy: predates the function-size gate; pending helper extraction" }
   }
 }
